+++ changelog.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.5.0
+
+### New features
+
+* \#112 Fire DOM change when choosing date
+
+* \#111 Remove container on input destroy
+
+
+### Resolved Issues
+
+* \#90 Today not respecting current timezone
+
+* \#87 ```getSelected``` does not return ```undefined``` anymore
+
+* \#87 Key presses in the input field no longer cause the field contents to be overwritten with an invalid date.
+
+
 ## 0.4.1
 
 ### Resolved Issues